src/main/webapp/app/entities/score-type/score-type-update.component.ts
import { Component, Vue, Inject } from 'vue-property-decorator';

import { required } from 'vuelidate/lib/validators';

import AlertService from '@/shared/alert/alert.service';

import { IScoreType, ScoreType } from '@/shared/model/score-type.model';
import ScoreTypeService from './score-type.service';

const validations: any = {
  scoreType: {
    name: {
      required,
    },
  },
};

@Component({
  validations,
})
export default class ScoreTypeUpdate extends Vue {
  @Inject('scoreTypeService') private scoreTypeService: () => ScoreTypeService;
  @Inject('alertService') private alertService: () => AlertService;

  public scoreType: IScoreType = new ScoreType();
  public isSaving = false;
  public currentLanguage = '';

  beforeRouteEnter(to, from, next) {
    next(vm => {
      if (to.params.scoreTypeId) {
        vm.retrieveScoreType(to.params.scoreTypeId);
      }
    });
  }

  created(): void {
    this.currentLanguage = this.$store.getters.currentLanguage;
    this.$store.watch(
      () => this.$store.getters.currentLanguage,
      () => {
        this.currentLanguage = this.$store.getters.currentLanguage;
      }
    );
  }

  public save(): void {
    this.isSaving = true;
    if (this.scoreType.id) {
      this.scoreTypeService()
        .update(this.scoreType)
        .then(param => {
          this.isSaving = false;
          this.$router.go(-1);
          const message = this.$t('xmplVueSpringwiseApp.scoreType.updated', { param: param.id });
          return (this.$root as any).$bvToast.toast(message.toString(), {
            toaster: 'b-toaster-top-center',
            title: 'Info',
            variant: 'info',
            solid: true,
            autoHideDelay: 5000,
          });
        })
        .catch(error => {
          this.isSaving = false;
          this.alertService().showHttpError(this, error.response);
        });
    } else {
      this.scoreTypeService()
        .create(this.scoreType)
        .then(param => {
          this.isSaving = false;
          this.$router.go(-1);
          const message = this.$t('xmplVueSpringwiseApp.scoreType.created', { param: param.id });
          (this.$root as any).$bvToast.toast(message.toString(), {
            toaster: 'b-toaster-top-center',
            title: 'Success',
            variant: 'success',
            solid: true,
            autoHideDelay: 5000,
          });
        })
        .catch(error => {
          this.isSaving = false;
          this.alertService().showHttpError(this, error.response);
        });
    }
  }

  public retrieveScoreType(scoreTypeId): void {
    this.scoreTypeService()
      .find(scoreTypeId)
      .then(res => {
        this.scoreType = res;
      })
      .catch(error => {
        this.alertService().showHttpError(this, error.response);
      });
  }

  public previousState(): void {
    this.$router.go(-1);
  }

  public initRelationships(): void {}
}
